Move pkg/slicer/manifest -> pkg/manifest
