+++ internal/pkg/manifest/manifest_test.go
