+++ internal/pkg/slicer/slice.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/keboola/processor-split-table/internal/pkg/config"
 	"github.com/keboola/processor-split-table/internal/pkg/log"
-	manifestPkg "github.com/keboola/processor-split-table/internal/pkg/slicer/manifest"
+	manifestPkg "github.com/keboola/processor-split-table/internal/pkg/manifest"
 	"github.com/keboola/processor-split-table/internal/pkg/slicer/rowsreader"
 	"github.com/keboola/processor-split-table/internal/pkg/slicer/slicedwriter"
 	"github.com/keboola/processor-split-table/internal/pkg/utils"
